feat(settings): 设置页 6 组重组 + 豆包统一配置卡真实转写测试
「测试连接绿灯、生产语音转文字不可用」的根因是链路错位：旧测试只用
0.1s 静音探针打文件极速版端点（auc_turbo），而实时录音走的是单独开通
的流式端点（sauc.duration）——测试结果对生产主路径零证明力。且三行
豆包凭据里只有 App ID 行接了测试函数，Token/Secret 行是空闭包死按钮；
Secret Key 全代码库零消费。

豆包侧：
- DoubaoConfigView 将 App ID + Access Token 合并为成对编辑的配置卡，
  删除 Secret Key 行（ApiKeyKind 保留 case 供 purge 覆盖）
- DoubaoConnectivityTester 用内置真人语音样本（DoubaoTestSample.wav，
  2.9s/16kHz，已对线上端点预验证）依次实测文件 + 流式两个生产端点，
  转写文字回显；断言用模糊匹配（去标点字符命中率 >=0.6）——流式灌音
  快于实时会丢中段词，全等匹配会误报
- 错误码翻译为中文指引（403 资源未开通/45000292 流式配额），验证态
  四态徽章（未配置/未验证/已验证+时间戳/未通过）持久化并逐级联动

设置页侧：
- 2055 行单文件拆为 hub + 7 个子页：账户/AI 与语音/提醒与通知/外观/
  同步与数据/关于；一级页每行带实时状态摘要
- 语音三处碎片（凭据/引擎选择/麦克风权限）合并进「AI 与语音」闭环
- 中英文各 +107 条本地化收口原硬编码中文；修掉给数字 App ID 显示
  "sk-..." 占位符的误导文案
- SettingsLabel 显式图标着色，根治 List 行复用丢 tint 的时棕时蓝
- 危险操作独立 Danger Zone；feature flags/调试事件流移入
  关于→开发者选项（Build 行连点 5 次解锁）

模拟器真实凭据实测：双端点全绿（文件 2.9s/流式 3.6s，转写正确）；
240 测试全通过（含 9 个新增单测）。
